Add Private ID DFCA spine combiner binary (#1824)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1824

# Context
**Internal:** [Private ID DFCA one-pager](https://docs.google.com/document/d/1mwImSRCljqU66uXWJS7jL-dQOrYAR4aBZ3bpJ8quiJw) Advertisers want to be able to target their customers on our platform in a privacy-safe manner. We utilise Private ID with the existing DFCA solution to provide this.

We're adding a new game type `PRIVATE_ID_DFCA` whose inputs for both parties are csvs with the columns `user_pii, user_id_{party}`, and the output is a csv of `publisher_user_id, partner_user_id` which can then be fed to our existing DFCA APIs.
After running the existing PID stages, we need to run two new binaries to get the output in its final form:
- private_id_dfca_spine_combiner
- private_id_dfca_aggregator

# This Diff
Adds the private_id_dfca_spine_combiner binary to the existing ID_SPINE_COMBINER stage.
The code is 99% the same as the PL/PA spine combiners.
We assume there are no duplicate PIIs in the data file and there is at most 1 match for every ID combination.

# This Stack
- **Add spine combiner binary**
- Add aggregation binary
- Add aggregation stage
- Add Stageflow
- Add game type

Reviewed By: jiancao-yajc

Differential Revision: D39926640

fbshipit-source-id: 760073973a954617f567273649ce9931ae437d6d