Improve E2E test run naming to make it easier for people to know what run is happening. (#1817)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1817

## Background
Right now the Actions page on FBPCS can be a little unhelpful to know what is actually happening for a E2E test run. We use the same action yaml for both single key an multikey test runs and so the names just show up as the same.

https://pxl.cl/2hRK7

In the case of a failure, you would need to dig deeper to know which run actually failed.

## This Diff
This diff adds the 'run-name' parameter, which is a way for us to dynamically name the run so it is more obvious which run is happening.

Reviewed By: joe1234wu

Differential Revision: D40687002

fbshipit-source-id: 76a7d0c9605111b810fd27d85c118191a93fad7d