Getting AWS region from the bucket name (#1841)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1841

**What**
Figuring out the AWS region within the AwsContainerLogs class from the input S3 bucket, suc that the caller no longer has to provide the AWS region input parameter.

**Why**
AWS region is required field to create boto3 clients. Till now we were deriving it from the container Id. Considering it's a required field we are improving to use a more reliable approach, in case there is no container ID provided in the input.

Reviewed By: gitfish77

Differential Revision: D40751409

fbshipit-source-id: f364df749ceba46b8f6b9c26f76a250f166f1267