check dummy rows when initializing InputData
Summary:
This diff made changes in initialization of `InputData` to identify dummy rows in `ID_spine`.

When reading `ID_spine` and converting it into `InputData`, we check if all timestamps in a row are all zeros. If so, we push a `true` into vector `isDummyRow_`. Otherwise, it's not dummy row and we push a `false`.

It is noted that a non-dummy row may have invalid timestamp. For example, a row may have a non-zero timestamp that is smaller than epic. We assume it is not a dummy row here.

Reviewed By: robotal

Differential Revision: D40709406

fbshipit-source-id: 9cabf6acc16a1f6a09ca2de3abe2b10539afc653