check dummy rows when initializing InputData (#1850)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1850

This diff made changes in initialization of `InputData` to identify dummy rows in `ID_spine`.

When reading `ID_spine` and converting it into `InputData`, we check if all timestamps in a row are all zeros. If so, we push a `true` into vector `isDummyRow_`. Otherwise, it's not dummy row and we push a `false`.

It is noted that a non-dummy row may have invalid timestamp. For example, a row may have a non-zero timestamp that is smaller than epic. We assume it is not a dummy row here.

Reviewed By: robotal

Differential Revision: D40709406

fbshipit-source-id: 8ae068444a165d50783e8f0bef6af84a3aebfb7a