Return appropriate provider class based on PC role
Summary:
When the feature gating for tls is on, we want to create the correct cert provider class depending on whether it's publisher or partner is running. For publisher side, we want to get PCInstance.*CertificateProvider because we will store certificate contents in pc instance on publisher side. For partner, we only need to get ca certificate, which will be retrieved from a parameter passed into the call. Currently, the certificate providers will return a static certificate

Next diffs:
1. implement `get_certificate()` in certificate providers properly, replacing the current logic of returning the content from some static files.

Reviewed By: danbunnell, musebc

Differential Revision: D40745837

fbshipit-source-id: 8418bbc49869b1a1fed95e6ce449db715a345b78