Get and store certificates when creating PCInstance (#1857)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1857

When creating a PCInstance for a study, we will:
1. get the certificate content
2. store them in the PCInstance Infra config for later consumption

This diff achieves #1 by reading cert contents from static files, which is an intermediary solution to be replaced by calling a CertificateCreator to generate certificate dynamically.

From this [note](https://fb.workplace.com/groups/126488202795965/permalink/401184761992973/), it is safe to add optional fields (fields with default values) in InfraConfig.

Next diff:
1. Implement the `get_certificate` logic in PCInstance.*CertificateProviders to read it from the infra config
2. Get and store server_hostname when creating PCinstance.

Reviewed By: joe1234wu

Differential Revision: D40816068

fbshipit-source-id: 308c66789742774844a7b208e1fdd579ac15b958