E2E test UDP input processor + PL logic correctness (#1832)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1832

With the `CompactionBasedInputProcessor` ready to substitute out the current `InputProcessor` implementation that is used for dummy UDP logic in milestone 2, we can update this E2E test to properly run the `DataProcessor` and `Adapter` flows within the PCF2 PL application.

This also strengthens the guarantee of the test added in D40402984 which only checks the processed inputs are present in the correct order when the intersection is equal to union. This test will generate a random data set with the possibility of some rows being dummy.

Currently the union map generation is a dummy implementation, so the intersection will always be equal to the union, but chennyc is working on  updating that.

Reviewed By: chennyc

Differential Revision: D40737248

fbshipit-source-id: 2d06b3a52867b6c043d1df4d83fe84beb32460b4