Modify synthetic data generator to simultaneously create partner and publisher files (#1856)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1856

There is a bug in the current synthetic data generator for PL stage which doesn't guarantee that one of the rows for publisher / partner will have a legit user (i.e. num_impressions > 1 or existing conversion value / timestamp). In the current lift stage this is ok since dummy values don't affect the end result, however with the addition of UDP we assume that each row will have at least one side matching, thus we can use `XNOR` in the adapter protocol. Since this assumption is not true in the synthetic data generator, it messes up the results and creates unmatched rows pointing to index 0. This will fix that by generating both files in unison.

Reviewed By: RuiyuZhu

Differential Revision: D40861009

fbshipit-source-id: 8b1d6b46feb4017079899c896ee05c19501b4670