Fix TARGETS files in //fbpcs/emp_games/lift (#1830)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1830

I am trying to add a test in PCF2_LIFT which depends on some of the newly added UDP code, however this is a bit difficult because none of the core business logic is wrapped in a `cpp_library` TARGET. This effectively means the same files have to be included in multiple build targets for `cpp_binary` or `cpp_unittest` targets.

I started making `cpp_library` modules that could be used across multiple projects and ended up with the following refactor. It also moves the `GenFakeData` tests to their own targets.

Reviewed By: haochenuw

Differential Revision: D40737249

fbshipit-source-id: c32602dd0035ddfc62c75334cfd6b9048bd5bff3