add server_uris to PCSMPCInstance
Summary: Each container should be assigned with a unique server uris, therefore adding this field in MPCInstance. This will be later used for configuring routing for the tls connection as certificates support for wildcard subdomain. We need a different address for every node because eventually we will expect the nodes to be looked up via DNS.

Differential Revision: https://internalfb.com/D40906402

fbshipit-source-id: 589613885c284df3c5761efba17bc006fc5ad134