Fix Scheduled E2E tests (#1870)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1870

## Background
Right now, the scheduled E2E test are failing because the input parameters that are required are not being passed to the actual test script. You can see the inputs being pased to the job:
https://pxl.cl/2jxc0
But when we try to run the `run_fbpcs.sh` script, none of the values are present:
https://pxl.cl/2jxcj

This is because the `workflow_call` syntax doesn't support the `github.event.inputs` context because the [`github.event` context](https://docs.github.com/en/actions/learn-github-actions/contexts#github-context) comes from the webhook payload and the [webhook payload for the `workload_call` event](https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#workflow_call) matches the payload of the caller. This means that the values that we pass in don't exist in the `github.event` context.

## This diff
Both the `workflow_call` and `workflow_dispatch` events support the [`inputs` context](https://docs.github.com/en/actions/learn-github-actions/contexts#inputs-context) so this diff updates the PA and PL E2E tests to use that context to get the input values.

Reviewed By: gorel

Differential Revision: D40939341

fbshipit-source-id: 88cd96d67aacd96914e9cbf83cb9a45cca13e456