add server_uris to PCSMPCInstance (#1872)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1872

Each container should be assigned with a unique server uris, therefore adding this field in MPCInstance. This will be later used for configuring routing for the tls connection as certificates support for wildcard subdomain. We need a different address for every node because eventually we will expect the nodes to be looked up via DNS.

Reviewed By: joe1234wu

Differential Revision: D40906402

fbshipit-source-id: 9fff7807dc95ee44097a883fdc0b174cf4329108