pass server_domain to infra config (#1899)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1899

Similar to D40816068 (https://github.com/facebookresearch/fbpcs/commit/94c3abb6569a5fdc6a15a67d90ddba43e8fa2b46), we want to store server_hostname on pc_instance to be sent later to the tls binary to set up the routing config for the tls connection at the time of creating a pc instance.

side note: added server_hostname to InfraConfig of PCInstance instead of as one of the fields in MPCInstance like ip_address is because 1. server_hostname is

Reviewed By: danbunnell, joe1234wu

Differential Revision: D40875886

fbshipit-source-id: 44f26c0a0cece3cc65eeea817150001b3fc99654