Inherit secrets into the One Command Runner and PA One Command runner workflows (#1903)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1903

## Background
The current E2E test workflow is failing on not bein able to reference the FBPCS_GRAPH_API_TOKEN. This is because called workflows don't automatically inherit secrets, so that value is empty when we pass the call to the One Command Runner and PA One Command Runner workflows. This causes the error seen below:
https://pxl.cl/2kcg2

## This diff
This diff passes the secrets through to the called workflows. The documentation for this is found [here](https://docs.github.com/en/actions/using-workflows/reusing-workflows#using-inputs-and-secrets-in-a-reusable-workflow).

Reviewed By: marksliva

Differential Revision: D41138443

fbshipit-source-id: 4443c54ccc97097df61bfc2d2fdc47081d259828