Temprorary rollback github config to unblock hotfix e2e test (#1918)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1918

## Why
In previous diff stack, we migrated mpc from fbpcp repo to fbpcs.
But during the code freeze, and hotfix we had, hotfix was based on the previous commit which didn't include the migration changes.
It only fail as we're duing the code freeze and RC is releasing with hotfix which isn't include the previous change.
So this this is to revert config in github as it's in-sync config, this change won't break any internal testing as we had fallback logic to compatible both class path
> fallback logic https://fburl.com/code/38u3ojo2
## What
- revert config been used in github back to fbpcp for temprorary

Reviewed By: musebc

Differential Revision: D41311591

fbshipit-source-id: c95eafd30cfea44f4a5f3c21c313a3ff097aaaa9