Update tests to run on GitHub Runners
Summary:
## Background
As a way to reduce our cloud cost, increase scalability, reduce complexity and improve our cloud security, I have started to move our self-hosted runners to use GitHub hosted runners. The first and easiest to migrate is the end-to-end test runners since they mainly just start a ECS task.

One of the problems we have with running our E2E test throughout the day is that we have a limited number of test runners and it could delay our regular release testing by blocking all the runners for an hour+ at a time.

## This diff
This diff updates the `run_fbpcs.sh` script to accept docker environment variables and pass them to the docker run command.

Differential Revision: https://internalfb.com/D40946624

fbshipit-source-id: f1c931809c63ed47613e6cba2b399580f0f0a046