Fix incorrectly labeled CMake option for EMP
Summary:
We've been adding this option to CMake called "EMP_USE_RANDOM_DEVICE" -- the intention is to *force* EMP to use `std::random_device` instead of `x86intrin.h` since the former is guaranteed to exist on all systems but the latter is CPU-dependent.

Unfortunately... we named it incorrectly. We shouldn't be using the `EMP_` prefix. By doing so, we were letting EMP decide whether to use `std::random_device` or `x86intrin.h` by checking if the build machine had access to `rdseed` internally.

What this means is that in some cases (like the ones described in the linked SEV task), we built on a machine with access to `rdseed`, but later when running in ECS, `rdseed` wasn't present. This gives the cryptic "Invalid Instruction" error because we're trying to execute a CPU instruction that doesn't exist on the machine where the computation is actually running.

Since there's no major difference to use between using `random_device` and `rdseed` and since we can't guarantee that we'll be given access to a machine with `rdseed` capabilities at runtime, we want to force EMP to always use `random_device`. This diff fixes that issue.

Differential Revision: D41193207

fbshipit-source-id: e3a60138b5f2ad33ded636a4811f7d26eaed2c5a