override default registries
Summary:
## What

- bolt_checkpoint set default registry value if given no key

## Why

Consider the following scenario:

1) Advertiser passes run id, meaning Bolt is NOT using a default run id
2) PCS code decorates a function with write_checkpoint that has no instance id arg (e.g. staticmethod)

This will result in the PCS code applying the default run id, NOT the same run id Bolt is using. With the updated implementation, Bolt will set the global default id if the advertiser passes a run id, ensuring a consistent run ID on the client side.

Bolt is in a good position to be the component setting the global defaults, as it is essentially the entry point to any private computation run.

## What is this stack

Add a decorator that can magically trace log with instance id and run id to the correct trace logger - no need to pass trace logging services all around the place:

```
bolt_checkpoint()
def my_func(...):
   ...

bolt_checkpoint(dump_params=True, dump_return_val=True)
def my_func(...):
   ...

# and much more ;)
```

I built the API so that it can be used in PCS as well, but I only added checkpointing in Bolt, since that is a mega gap right now.

{F802371475}

Reviewed By: joe1234wu

Differential Revision:
D41465180

LaMa Project: L416713

fbshipit-source-id: a0e06ada551284572c8447f8513627248413f388