Add crypto key components for use in TLS (#1930)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1930

We are adding support for TLS in Private Lift. In a subsequent change, we will need to generate TLS certificates. To generate a certificate, cryptographic keys will be used for signing.

This change adds support for generating and using Elliptic-Curve Cryptography keys for digital signatures. This will later be used during certificate generation.

We will be using ECC keys because they can achieve high security with a short key length: 244-bit ECC is similar to 2048-bit RSA. ECC is faster for key generation and signing, but RSA is faster for signature verification.

The module was designed to abstract the underlying implementation of the cryptography, which relies on the widely used PYCA Cryptography library. It also hides most of the crypto choices from the consumer, to avoid misconfiguration.

Reviewed By: Kewen12

Differential Revision: D41269734

fbshipit-source-id: 95f49aeb4470dad8f026d77935d9c8cd74f693fb