fix a flaky test_create_instance (#1974)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1974

This diff fixes a flaky test `test_create_instance`, which failed occasionally (P565700528, P565706383).

In this test, there is a check to ensure the correctness of `dataset_timestamp`:
```
self.assertAlmostEqual(
                    int(yesterday_timestamp),
                    args.product_config.common.post_processing_data.dataset_timestamp,
                )
```
It's possible that the truth `yesterday_timestamp` is larger than `dataset_timestamp` by 1 second because these two timestamps were not generated at the same time and rounded to different values, which will fail this test.

This diff fixes this by adding a delta `1` second as a parameter.

Reviewed By: RuiyuZhu

Differential Revision:
D41700685

LaMa Project: L416713

fbshipit-source-id: aaabf70368bc4c65a9b7abdaa752b329868ccd3e