Add feature flags for PA and PL timestamp validation (#1991)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1991

In order to control the release process for the timestamp validations feature,
add 2 new feature flags so that it can be rolled out smoothly. These flags will
also enable us to quickly roll back the validations if needed.

For now, these feature flags are not returned from the Graph API. So they
will default to false in the meantime and empty timestamps will be sent to
the pre_validation CLI which makes it skip the timestamp validation.

In a followup after the code freeze, the new feature flags will be added into
the `www` layer so that this validation can be released.

Reviewed By: joe1234wu

Differential Revision: D41981227

fbshipit-source-id: db290c9c39af58baaf04d4e6f3218dd8d7e2c11a