Change container layer interface to accept list of env_vars (#2008)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2008

X-link: https://github.com/facebookresearch/fbpcp/pull/475

One of the requirements of TLS/OPA integration is to be able to send a set of env_vars for each container. Currently, the infra only supports a single set of env_vars for all containers. This stack of diffs is to enable this feature for private lift runs.

This diff:
* modify the container interface to accept a list of env_vars in addition to a single env_vars dict in `create_instances`
* this diff only changes the interface and the functionality remains the same.

Next:
* changes in onedocker layer
* changes in pcs layer (note that this would have fbpcp release dependency)

Reviewed By: liliarizona

Differential Revision: D42199006

fbshipit-source-id: d32eb53abba97e2027d3556920480d4b03bb09e6