Fix a Flaky Test in CalculatorAppTest (#1919)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1919

## What
Modified CalculatorGame's `playFromSecretShare` method to return a dummy item with the publisherBreakdown field populated with dummy items as well, when the number of rows is zero.

Added test cases for empty intersection.

Added "long_running" and "heavyweight" tags to the test target, to reduce the probability of tests timing out.

## Why
A Calculator app test with random input shows as flaky. After investigation, it seems that the test fails when 1) UDP is turned on and 2) the intersection size = 0. In this case, both expected and actual results are dummy, but their formats are different, causing the test to fail.

The intersection size = 0 happens with a small probability, which is why the test only fails sometimes.

There are also some tests in the suite that requires longer than 600 seconds to run. So I added the tags to increase the timeout limit.

## Investigation process
https://docs.google.com/document/d/1raKj9_aaXUED5oCc48ICc5qyf_Anp0Fcy79ZhzV_YTM/edit#

Reviewed By: robotal

Differential Revision: D41241066

fbshipit-source-id: ba0ecbba728c8f357ab19dfd534bca09796d191e