Update PCS create instance lib to take private key secret reference as parameter
Summary:
This change updates the PCS library for creating a PCS instance to accept a parameter referring to the server private key. The private key reference will later be passed to the Publisher containers so they can access the key and authenticate for TLS during MPC.

This feature will be disabled by a feature flag and generally in production until manually tested e2e.

Reviewed By: liliarizona

Differential Revision: D42008712

fbshipit-source-id: 679d0f5b09b65ce7103a282672b6a65855c8f611