Open Source the production OneDocker Dockerfile. (#1924)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1924

## Context
We are in the process of trying to both release the OneDocker image as part of our normal build and release process, and publish the image to a new [public ECR repository](https://docs.google.com/document/d/1XDnnMCQ6IhyKq9VUhaxat8ns5ddyBqGQSgpbubitpz8/edit?usp=sharing). This stack of changes will set up that process so in 2023 we can move to weekly releases of the production OneDocker image.

## This diff
This diff updates the production OneDocker image to be open sourced as part of the FBPCS Repository.

Reviewed By: marksliva

Differential Revision: D41227167

fbshipit-source-id: 682093e15b825e145bb0690cc2d625c5edb8a477