Open Source the production OneDocker Dockerfile. (#1924)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1924

## Context
We are in the process of trying to both release the OneDocker image as part of our normal build and release process, and publish the image to a new [public ECR repository](https://docs.google.com/document/d/1XDnnMCQ6IhyKq9VUhaxat8ns5ddyBqGQSgpbubitpz8/edit?usp=sharing). This stack of changes will set up that process so in 2023 we can move to weekly releases of the production OneDocker image.

## This diff
This diff updates the production OneDocker image to be open sourced as part of the FBPCS Repository.

Differential Revision: https://internalfb.com/D41227167

fbshipit-source-id: 62ed63389112fe5aae8d2e4a5c7237cdb8b98af0