Publish the production OneDocker images on pushes to main that affect the files included in it to the Github Container Registry (#1960)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/1960

## Context
We are in the process of trying to both release the OneDocker image as part of our normal build and release process, and publish the image to a new [public ECR repository](https://docs.google.com/document/d/1XDnnMCQ6IhyKq9VUhaxat8ns5ddyBqGQSgpbubitpz8/edit?usp=sharing). This stack of changes will set up that process so in 2023 we can move to weekly releases of the production onedocker image.

## This Diff
This diff publishes the production OneDocker Image to the GitHub Container registry so we can start testing with it in ECS.

Reviewed By: joe1234wu

Differential Revision: D41228111

fbshipit-source-id: f3853a98412a21b772360f9da1fe3247dcb07bb9