Add argument validation for permutations and unionMaps (#458)
Summary:
In T139445158 the NCC brought to our attention that the permuter library does not actually verify that the user has provided a valid permutation (i.e. same size as dataset, each index from [0 to n-1] appears exactly once. This diff adds that input validation.

The other validation that is added is in the adapter is to check for a valid union map. A valid union map must have index that range from 0 -> `n-1` or -1. Furthermore each union index can only appear at most once.

Note: The other party that calls permute or adapt will just hang and timeout when this happens as it has no way to know the other party aborted.

X-link: https://github.com/facebookresearch/fbpcf/pull/458

Differential Revision: D42014843

fbshipit-source-id: dddc43bf73f33b74122e3e788709053810a4626d