Increase Computation timeout from 12 hours to 24 hours. (#2033)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2033

## Context

For a run of large cohorts in January, we are temporarily increasing the lift timeout to 24 hours from 12 hours. This has some risk that we could burn through extra cash if we run into the situation of partner pods failing to connect and containers just wait around for a connection that will never happen. T142284889 was created to revert this change as soon as we are done with the current run and have a better solution.

Reviewed By: jrodal98

Differential Revision: D42479105

fbshipit-source-id: 58e0875641ea595474df042c26cc8c93ade6e169