Pass server private key reference to Publisher containers during deployment (#2037)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2037

This change updates PCS stages to pass a Private Key reference provider when running a stage. The provider is used to generate environment variables for the container which describe how to access the Server Private Key for TLS.

PCS will only generate these environment variables on the Publisher (server) side, when the TLS feature is enabled, and the required data is available on the PC Instance.

Reviewed By: joe1234wu

Differential Revision: D42480836

fbshipit-source-id: a7e7ec4843282e65051a8319c26e7567da7ff241