Pass env_var per container in PCS layer [BLOCK LANDING TILL FBPCP RELEASE] (#2043)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2043

We have changed the container and onedocker interface and implementation in previous diffs in the stack. Our goal is to be able to pass env_vars per container from PCS because of the requirement from enabling TLS, hence this change.

Note that the shipping of this diff will be blocked by releasing new fbpcp pypi package with the changes in this diff stack: D42211965

Reviewed By: joe1234wu

Differential Revision: D42331259

fbshipit-source-id: ac3dbde526320386afec0b2e02c2d3f1cf6e04da