Update the EMP Games and Data Processing builds to use all the cores on the EC2 instance (#2048)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2048

## Summary
While working on T133501570 I noticed that we were maxing out a single core for an extended time while building the FBPCF image. This led me to investigate our other builds and I noticed that we were doing the same here. The EC2 instance we use to build has 16 cores and we were only using one, leading us to only use ~6.5% of the available CPU. By increasing the number of make jobs to match the number of cores, I was able to increase the amount of CPU used to ~23% and reduce the build time from an average of around 40 minutes to ~19 minutes.

**Memory Per Thread Selection**
I spent some time running the builds on different build machines and monitoring the memory usage of the individual build jobs. I found that the maximum memory used by one of the FBPCS processes was ~7GBs. This means that we should allocate that much per process since we will need at least that much available to keep the processes from moving into swap and slowing to a crawl until the OOM killer comes along and ends one for us. All this does is reduce the number of jobs run on a machine based on available memory. For example, on our GitHub runners, we have 2 vCPUs and 7GB of memory. Just based on the number of cores, we should be able to run 2 jobs at once, but as I found out with early versions of this diff, that won't work as they will get locked in the amount of memory available and eventually fail because the OOM killer will kill something early.

{F855357152}

Differential Revision: D42626512

fbshipit-source-id: dd82062ce23eb1b694921c8ff6a6d181d862d183