Temporarily disable the E2E cron tests and migrate all objective IDs back to the release blocking tests. (#2059)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2059

## Context
This should be a work-around for T143330706 until we come up with a better solution to select unused objective IDs for the E2E tests (T143464466).

## This diff
This diff returns the extra objective IDs that were being used for the cron tests to the nightly tests. This should alieviate the problems of T143330706 for the Multi-Key PL tests. We may still have issues with the RC tests.

Reviewed By: gitfish77

Differential Revision: D42744036

fbshipit-source-id: 69b6493cdc20415f873ed3a9720f441f186776ba