Update Binary image cache to key on branch instead of hash
Summary:
## Context
The binary build caches have been keyed on the commit hash. This means that we were creating a new Docker Builder cache for each commit on the main branch. This quickly fills up our alloted 10GB cache.

## This Diff
This change updates the cache key to be based on the `ref_name` which is the branch or tag the job is running on ([docs](https://docs.github.com/en/actions/learn-github-actions/contexts#github-context)). This will create a single cache for the main branch. This will not affect the builds at all as the cache is based on a hash of the files so any changes will trigger a full rebuild of that layer and any subsequent layers anyway.

Differential Revision: D42847337

fbshipit-source-id: eac998eb4f577b8fc5608fcec09cededa4db3c4e