Update FBPCS Binaries to be Built with the Haswell Architecture (#2066)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2066

## Context
After attempting to build the images with the "x86-64" architecture for a while, I determined that we wouldn't be able to make it work as the EMP library used an instruction that wasing included in the "x86-64" architecture (specifically ). This meant that I needed to explore other options. My next choice was to find the architecture of the current build machines that we use (c4.4xlarge) and go with that since we know that we were building it previously natively for those machines. The c4.4xlarge uses an Intel Xeon E5-2666 based on the cpu info from the machine:
```
ubuntu@ip-172-31-15-42:~$ lscpu
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
...
Model name:                      Intel(R) Xeon(R) CPU E5-2666 v3 @ 2.90GHz
...
```

According to the Intel website, this processor uses the Haswell architecture: https://www.intel.com/content/www/us/en/products/sku/81706/intel-xeon-processor-e52660-v3-25m-cache-2-60-ghz/specifications.html

With that information, I decided to rebuild everything for the haswell architecture since it should match our current build system and allow us all the symbols necessary to run the computation. This means that this isn't really an architecture change as much as making the current architecture explicit in out builds, allowing us to use other build machines without including illegal instructions.

## This Diff
This diff updates the FBPCS binaries to specify the C and C++ flags to build with the haswell architecture.

Reviewed By: calebzj

Differential Revision: D42713487

fbshipit-source-id: 35b7004726e571a807a948be3a6d2012c32c8b79