Adding validation for purchase value in pc_pre_validation
Summary:
HG: Context
We recently faced a SEV with Amazon, where the computation stage failed because of an invalid (greater than INT.MAX) purchase value input. This should validated at the beginning on the run in PC_PRE_VALIDATION stage.

HG: Changes
In this diff, adding a validation for the purchase value, with the valid range being 0 to 2147483647. This will help us detect similar issue in the beginning and avoid another SEV.

Differential Revision: D42945492

fbshipit-source-id: 051fd4782ed4b4a87a306e6fa2c66022ef99d211