Adding validation for purchase value in pc_pre_validation (#2072)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2072

# Context
We recently faced a SEV with Amazon, where the computation stage failed because of an invalid (greater than INT.MAX) purchase value input. This should validated at the beginning on the run in PC_PRE_VALIDATION stage.

# Changes
In this diff, adding a validation for the purchase value, with the valid range being 0 to 2147483647. This will help us detect similar issue in the beginning and avoid another SEV.

Reviewed By: yanglu-fb, marksliva

Differential Revision: D42945492

fbshipit-source-id: 0daae2cd223c36dd5901817b207e89e75aa14b09