Gradually remove 'usingBatch' - attribution game and metrics
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution game and attribution metrics class

Differential Revision: https://internalfb.com/D43103054

fbshipit-source-id: 642f85a8809a862816e11634b91f82751d70c064