Gradually remove 'usingBatch' - attribution rule
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution rule object

Differential Revision: https://internalfb.com/D43103055

fbshipit-source-id: 3cbd2ef74555cf7eac0029f052ccef84cd12044c