Gradually remove 'usingBatch' - conversion
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the conversion-related objects

Reviewed By: zhangpuhan

Differential Revision: D43103048

fbshipit-source-id: fb8eaca9049b212a75ac8353cc63b99e4cb4c62f