Gradually remove 'usingBatch' - touchpoint
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. In this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the touchpoint related objects

Reviewed By: zhangpuhan

Differential Revision: D43103051

fbshipit-source-id: 3f0960b49cf7946a65cb60449babdbf3b7a4a780