Gradually remove 'usingBatch' - attribution reformatted output
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the reformatted output object.

Differential Revision: https://internalfb.com/D43103046

fbshipit-source-id: d435c0d8cf1e0f37f4c2cc0994afa592e36c0a1c