Gradually remove 'usingBatch' - attribution output
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution output object

Differential Revision: https://internalfb.com/D43103047

fbshipit-source-id: 1ce097d366b091553924bc3ad95937935c890528