Gradually remove 'usingBatch' - touchpoint
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the touchpoint related objects

Differential Revision: https://internalfb.com/D43103051

fbshipit-source-id: 50a121032e58b7fc489384261f1c9a1be9f77fff