Gradually remove 'usingBatch' - conversion
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the conversion-related objects

Differential Revision: https://internalfb.com/D43103048

fbshipit-source-id: c1164aa0f5f0febc4ad1c623f8eea8996f8a5e85