Gradually remove 'usingBatch' - attribution rule
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution rule object

Differential Revision: https://internalfb.com/D43103055

fbshipit-source-id: 49364645b818ff4e8a26e174991656316c73c4fe