Gradually remove 'usingBatch' - attribution rule
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution rule object

Reviewed By: zhangpuhan

Differential Revision: D43103055

fbshipit-source-id: b4290580554a3c0ba659325b3ce7ec560a1e43f2