Gradually remove 'usingBatch' - constant
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. In this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the constants and basic types at the lowest level and complete the code clean-up

Reviewed By: zhangpuhan

Differential Revision: D43103052

fbshipit-source-id: 6d533e0c7b4566bdb416a3e60cec2ec27bb7251b