Gradually remove 'usingBatch' - attribution reformatted output
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the reformatted output object.

Reviewed By: zhangpuhan

Differential Revision: D43103046

fbshipit-source-id: d6c3c7f0ce0df98ffbf1075afdcb6634ac9f89c2