Gradually remove 'usingBatch' - attribution game and metrics
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution game and attribution metrics class

Reviewed By: zhangpuhan

Differential Revision: D43103054

fbshipit-source-id: b06c172f6ad7a5d89cae3e8f75f87e8c8e20ea87