Gradually remove 'usingBatch' - attribution app
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to attribution app class

Reviewed By: haochenuw

Differential Revision: D43103050

fbshipit-source-id: b20af152dfe2eaa3a5c7bb65fe6c1d819814bfbd