Gradually remove 'usingBatch' - main files
Summary:
Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the main file.

Reviewed By: anthonyzhang25

Differential Revision:
D43103045

Privacy Context Container: L416713

fbshipit-source-id: 005a387dccc26e521208ca11c8ef353f6bd57169