Gradually remove 'usingBatch' - touchpoint (#2098)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2098

Context: in attribution games, we always use batch mode for the sake of performance. In this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the touchpoint related objects

Reviewed By: zhangpuhan

Differential Revision: D43103051

fbshipit-source-id: bc358076f5cf8046aac56e6256ee2d25bf10350a