Gradually remove 'usingBatch' - attribution reformatted output (#2100)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2100

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the reformatted output object.

Reviewed By: zhangpuhan

Differential Revision: D43103046

fbshipit-source-id: 837cd6a19cc1abbc7d55d1c084b187c0c007ae45