Gradually remove 'usingBatch' - attribution app (#2101)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2101

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to attribution app class

Reviewed By: haochenuw

Differential Revision: D43103050

fbshipit-source-id: b5b3e6472a1c5c836b7936b5a4ae2b8d3b99ef6e