Gradually remove 'usingBatch' - attribution game and metrics (#2103)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2103

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution game and attribution metrics class

Reviewed By: zhangpuhan

Differential Revision: D43103054

fbshipit-source-id: a1473bcbfcd556390684dfedb3f24766f6722cf8