Gradually remove 'usingBatch' - main files (#2105)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2105

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the main file.

Reviewed By: anthonyzhang25

Differential Revision:
D43103045

Privacy Context Container: L416713

fbshipit-source-id: ce566c32a1a05ca622b6e507405962cae7266560