Gradually remove 'usingBatch' - conversion (#2097)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2097

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the conversion-related objects

Reviewed By: zhangpuhan

Differential Revision: D43103048

fbshipit-source-id: e497792f94bb35d1ebbdf1b4d761230f7905880d