Gradually remove 'usingBatch' - attribution rule (#2099)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2099

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution rule object

Reviewed By: zhangpuhan

Differential Revision: D43103055

fbshipit-source-id: 57e7c623e5247ee61ccf52d1da336e31eff63146