Gradually remove 'usingBatch' - attribution output (#2102)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2102

Context: in attribution games, we always use batch mode for the sake of performance. IN this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the attribution output object

Reviewed By: zhangpuhan

Differential Revision: D43103047

fbshipit-source-id: d5f0f84355f77e4aab9c3cc13ae023e3c102e95e