Gradually remove 'usingBatch' - constant (#2104)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2104

Context: in attribution games, we always use batch mode for the sake of performance. In this stack of diffs, we are gradually remove the code for non-batch path and hardcode usingbatch = true.

We are taking a top-down approach to delete this flag layer by layer.

This diff contains the change to the constants and basic types at the lowest level and complete the code clean-up

Reviewed By: zhangpuhan

Differential Revision: D43103052

fbshipit-source-id: 5b08e3fdc9b6eeb04798869c37bc35f731a4e489