add identifier filter in PID PREPARE (#2058)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2058

# Context
Through pilot runs, we found that we have an issue of low quality identifiers for both publisher-side and partner-side. Therefore, we decided to add a change into PID PREPARE stage to have ability to filter out identifiers based on the number of appearances.

# Details
In order to achieve identifier filtering based on the number of appearances, we are additionally reading through an input file.
- Added new arg `id_filter_thresh`, which sets the threshold for the number of appearances allowed. The default value is -1m and it does not apply filtering by default.
- Added another file reader `bufferedReaderForFilter` to read through a file. It counts appearance of each identifier and store ones that exceeded `id_filter_thresh` in `filterIds`.
- The rest of the process is the same except that it drops an identifier if an identifier is in `filterIds`.
- Added a new unit test `AttributionIdSpineInputValidationWithFilter`. It makes sure the identifier is being filtered as expected.

Reviewed By: Pradeepnitw

Differential Revision: D42697977

fbshipit-source-id: cab2fecfce84169ad679629270cf7528f822ef53