pass identifier filter to PID PREPARE (#2068)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2068

# Context
Through pilot runs, we found that we have an issue of low quality identifiers for both publisher-side and partner-side. Therefore, we decided to add a change into PID PREPARE stage to have ability to filter out identifiers based on the number of appearances. For overall context, please visit [the doc](https://docs.google.com/document/d/1Zr34S5AjOSt_Kjt-dX2-dllypfImQPLN600ycNJrfQM/edit?usp=sharing).

# Details
We have added a PCS feature `pid_filter_low_quality_identifier_thresh166`. When the feature is turned on, we will be sending the threshold of 166. Otherwise, we will be sending the threshold of -1 to do no filtering.

Reviewed By: rajprateek

Differential Revision: D42854941

fbshipit-source-id: 65160f8af43649a36696b337cdfe94f1df002349