remove encryption method in attribution app and main
Summary:
In this stack of diff, we will gradually remove `inputEncyption` variable in the template as this should be a run time parameter.

This time we take a bottom-up approach to gradually propagate the changes upwards.

This diff changes the attribution app and main function

Differential Revision:
D43261903

Privacy Context Container: L416713

fbshipit-source-id: 4c08bfc767055ea558c11d5aee2dbeff0d6009c0