Remove encryption method from touchpoints
Summary:
In this stack of diff, we will gradually remove `inputEncyption` variable in the template as this should be a run time parameter.

This time we take a bottom-up approach to gradually propagate the changes upwards.

This diff make the changes related to touch points

Differential Revision:
D43257626

Privacy Context Container: L416713

fbshipit-source-id: e882915ff2b8f5b28f64f1c8ce7625fa923cd0f1