refactor the util functions to support passing in a constructor (#2115)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2115

In this stack of diff, we will gradually remove `inputEncyption` variable in the template as this should be a run time parameter.

This time we take a bottom-up approach to gradually propagate the changes upwards.

This diff will refactor the util functions to prepare for the change.

Reviewed By: anthonyzhang25, chennyc

Differential Revision:
D43256421

Privacy Context Container: L416713

fbshipit-source-id: fe97ec4282b79cb8b869bf1d74dd11532ae3aa69