remove encryption method in attribution app and main (#2113)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2113

In this stack of diff, we will gradually remove `inputEncyption` variable in the template as this should be a run time parameter.

This time we take a bottom-up approach to gradually propagate the changes upwards.

This diff changes the attribution app and main function

Reviewed By: chennyc

Differential Revision:
D43261903

Privacy Context Container: L416713

fbshipit-source-id: c7e91c59bc5159aa703f1888b7a384acf791c893