Remove encryption method from touchpoints (#2114)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2114

In this stack of diff, we will gradually remove `inputEncyption` variable in the template as this should be a run time parameter.

This time we take a bottom-up approach to gradually propagate the changes upwards.

This diff make the changes related to touch points

Reviewed By: chennyc

Differential Revision:
D43257626

Privacy Context Container: L416713

fbshipit-source-id: 4834a870eea971dc848f158e8fa93f5de7b2b483