Add logging for number of id columns
Summary:
## Why
The purpose is to log how many multi-key the advertisers are using.
In PID SHARD stage, we have "rowsInShard" to log the number of rows of each shard. And now we need to log the number of columns of the file. So we rename "rowsInShard" to be more generic "pid_shard_info".
## What
 read the file header to calculate the number of columns;

Reviewed By: yuyashiraki

Differential Revision: D43066928

fbshipit-source-id: c7f53a97e850f7336aac2624580fabfd1e5a9b06