Add PostUDPInputProcessor (#2132)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2132

What: added a stub UDPBasedInputProcessor, which will be called in the beginning of Lift stage to directly work on serialized secret shares produced by UDP decryption (as a pair of double nested vector of bools), and produce the LiftGameProcessedData.

NOTE: The actual functionality will be completed in the following diffs in the same stack.

Why: this is part of the UDP upgrade + UDP for PA project to upgrade UDP to be product agnostic.

Reviewed By: RuiyuZhu

Differential Revision:
D43256659

Privacy Context Container: L1143502

fbshipit-source-id: 9c37404a485b94acbaddd2f2e9d745a08c420c7d