Crypto utility for encryption & decryption (#2126)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2126

# Why
This library will be used to encrypt / decrypt shares from ad-hoc advertisers.

# What
The diff implements a light wrapper on OpenSSL hybrid encryption/decryption functions, based on RSA (for encrypting the session key) and AES_256_CBC (for encrypting the actual data).

TODO (in subsequent diffs):
* Add signatures to cipher message (authentication).
* Add serialization / deserialization.

# Risk
Unfortunately, the API of OpenSSL libcrypto is a bit messy, and they are constantly deprecating mysterious APIs in newer versions. So we may have to come back to update the code.

We have tried to use only APIs that are not deprecated in OpenSSL 3.1.

Reviewed By: mdas7, haochenuw

Differential Revision: D43255201

fbshipit-source-id: a7a8ebaefd3b39daf4645425d54be80bb351cc17