enhanced print_log_urls cli-endpoint allow partner can retreive all stages or failed only cw logs (#2159)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2159

## Why
In recent sev S321509, as a sev follow up task, we want to provide a better way that advertiser can locate the failed cw logs.
The previous ask of the task is to add the runbook, But I think to enhanced our existing `print_log_urls` cli-endpoint would be easier for partner to locate.

## What
- In current `print_log_urls`, it's only print the recently run stage cw logs.
- adding two optional argument for print_log_urls. `--all_stages` and `--failed_only`
 - `--all_stages` would iterate all stages from instance config
 - `--failed_only` would only return failed containers logs. for easier locate the problematic log we're more concerning

Reviewed By: jrodal98

Differential Revision: D43931705

fbshipit-source-id: 3bcae87b67696bb7a70de61baa39edf2bb4687e1