Call wait_valid_publisher_status before initiating a retry.
Summary:
# Context
In co-ordinated retry, during partner failure in joint stage, we cancel the partner and publisher stage containers in order to restart the stage.
In some instances we have noticed that even after the wait period to retry, when we poll the publisher stage, we get a FAILED status (publisher transitions from PROCESSING_REQUEST to FAILED due to the cancel_stage).

This is causing us to lose one retry.

# Changes in diff
In order to fix this issue, in this diff. Adding a call to wait for valid publisher status before we trigger a retry.

Differential Revision: D44152010

fbshipit-source-id: c866bdc1042dae0e7e071bf759e63ae15e8ba470