Call wait_valid_publisher_status before initiating a retry. (#2194)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2194

# Context
In co-ordinated retry, during partner failure in joint stage, we cancel the partner and publisher stage containers in order to restart the stage.
In some instances we have noticed that even after the wait period to retry, when we poll the publisher stage, we get a FAILED status (publisher transitions from PROCESSING_REQUEST to FAILED due to the cancel_stage).

This is causing us to lose one retry.

# Changes in diff
In order to fix this issue, in this diff. Adding a call to wait for valid publisher status before we trigger a retry.

Reviewed By: jrodal98

Differential Revision:
D44152010

Privacy Context Container: L416713

fbshipit-source-id: 2f4d92217af82aba737dd8edffb8b36621aa8476