Pass OPA workflow path from PCS
Summary: The TLS feature has integrated with OneDocker plugin architecture to run the TLS script inside onedocker containers. We want to pass the opa_workflow_path command from PCS to tell OPA which plugin workflow we want to run. The workflow json files will be embedded in the onedocker image. If no path is provided, no plugin will be triggered, and everything is supposed to be the same as before.

Reviewed By: joe1234wu

Differential Revision: D43998127

fbshipit-source-id: c38c13c54011f354f300fda1267422283238e104