Testing the HE Aggregation Game
Summary:
This diff adds a correctness test to the HE aggregation game.

The test contain 4 input data files:

<ss_publisher_0.csv> contains publisher's secret share attribution result
<ss_partner_0.csv> contains partner's secret share attribution result
<dataproc_publisher_0.csv> contains publisher's touchpoint information (i.e., aggregation key such as  ad_id in this case)
<dataproc_partner_0.csv> contains partner's conversion information

The diff contains a new test which reads the input files, initializes the Aggregation Input class and calls the HEAggGame to compare the result with the expected result.

Reviewed By: mdas7

Differential Revision:
D43977171

Privacy Context Container: L1173372

fbshipit-source-id: b612a487066ee8affdba07d118705d00b35d869a