Testing the HE Aggregation Game (#2228)
Summary:
Pull Request resolved: https://github.com/facebookresearch/fbpcs/pull/2228

This diff adds a correctness test to the HE aggregation game.

The test contain 4 input data files:

<ss_publisher_0.csv> contains publisher's secret share attribution result
<ss_partner_0.csv> contains partner's secret share attribution result
<dataproc_publisher_0.csv> contains publisher's touchpoint information (i.e., aggregation key such as  ad_id in this case)
<dataproc_partner_0.csv> contains partner's conversion information

The diff contains a new test which reads the input files, initializes the Aggregation Input class and calls the HEAggGame to compare the result with the expected result.

Reviewed By: mdas7

Differential Revision:
D43977171

Privacy Context Container: L1173372

fbshipit-source-id: f46efc3d8f886910d85d982710cdcda314d238db